+++ .idea/misc.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Black">
+    <option name="sdkName" value="Python 3.12.3 WSL (Ubuntu): (/home/bouma/projects/mcp-tools/gcp-mcp/venv/bin/python3.12)" />
+  </component>
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.12.3 WSL (Ubuntu): (/home/bouma/projects/mcp-tools/gcp-mcp/venv/bin/python3.12)" project-jdk-type="Python SDK" />
+</project>
