+++ requirements.txt
@@ -1,10 +1,67 @@
-google-cloud-storage>=2.0.0
-google-cloud-compute>=1.0.0
-google-cloud-bigquery>=3.0.0
-google-cloud-pubsub>=2.0.0
-google-cloud-secret-manager>=2.0.0
-google-cloud-monitoring>=2.0.0
-google-cloud-billing>=1.0.0
-python-dotenv>=0.19.0
-fastmcp>=1.0.0
-pydantic>=1.8.0
+annotated-types==0.7.0
+anyio==4.9.0
+Authlib==1.6.0
+cachetools==5.5.2
+certifi==2025.6.15
+cffi==1.17.1
+charset-normalizer==3.4.2
+click==8.2.1
+cryptography==45.0.4
+exceptiongroup==1.3.0
+fastmcp==2.9.2
+google-api-core==2.25.1
+google-auth==2.40.3
+google-cloud-bigquery==3.34.0
+google-cloud-billing==1.16.3
+google-cloud-compute==1.31.0
+google-cloud-core==2.4.3
+google-cloud-monitoring==2.27.2
+google-cloud-pubsub==2.30.0
+google-cloud-secret-manager==2.24.0
+google-cloud-storage==3.1.1
+google-crc32c==1.7.1
+google-resumable-media==2.7.2
+googleapis-common-protos==1.70.0
+grpc-google-iam-v1==0.14.2
+grpcio==1.73.1
+grpcio-status==1.73.1
+h11==0.16.0
+httpcore==1.0.9
+httpx==0.28.1
+httpx-sse==0.4.1
+idna==3.10
+importlib_metadata==8.7.0
+markdown-it-py==3.0.0
+mcp==1.9.4
+mdurl==0.1.2
+openapi-pydantic==0.5.1
+opentelemetry-api==1.34.1
+opentelemetry-sdk==1.34.1
+opentelemetry-semantic-conventions==0.55b1
+packaging==25.0
+proto-plus==1.26.1
+protobuf==6.31.1
+pyasn1==0.6.1
+pyasn1_modules==0.4.2
+pycparser==2.22
+pydantic==2.11.7
+pydantic-settings==2.10.1
+pydantic_core==2.33.2
+Pygments==2.19.2
+python-dateutil==2.9.0.post0
+python-dotenv==1.1.1
+python-multipart==0.0.20
+requests==2.32.4
+rich==14.0.0
+rsa==4.9.1
+shellingham==1.5.4
+six==1.17.0
+sniffio==1.3.1
+sse-starlette==2.3.6
+starlette==0.47.1
+typer==0.16.0
+typing-inspection==0.4.1
+typing_extensions==4.14.0
+urllib3==2.5.0
+uvicorn==0.35.0
+zipp==3.23.0
